requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
fastapi-users
fastapi-users-db-sqlalchemy
greenlet
h11
idna
iniconfig
makefun
packaging
passlib
pluggy
psycopg2-binary
pyasn1
pycodestyle
pycparser
pydantic
PyJWT
pytest
python-dotenv
python-jose
python-multipart
requests
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
emoji
pandas
bumpversion
pdfkit 
jinja2
inflect
redis
openpyxl
ldap3
requests-ntlm
opencv-contrib-python-headless
pytz
Pillow
openai